initial typescript porting
