+++ MotorVehicleManagerSystem/.gitignore
@@ -1,3 +1,17 @@
-node_modules/**/*
-.expo/*
-npm-debug.*
+# See https://help.github.com/ignore-files/ for more about ignoring files.
+
+# expo
+.expo/
+
+# dependencies
+/node_modules
+
+# misc
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
