+++ MotorVehicleManagerSystem/app.json
@@ -1,27 +1,5 @@
 {
   "expo": {
-    "name": "MotorVehicleManagerSystem",
-    "description": "This project is really great.",
-    "slug": "MotorVehicleManagerSystem",
-    "privacy": "public",
-    "sdkVersion": "29.0.0",
-    "platforms": ["ios", "android"],
-    "version": "1.0.0",
-    "orientation": "portrait",
-    "icon": "./assets/icon.png",
-    "splash": {
-      "image": "./assets/splash.png",
-      "resizeMode": "contain",
-      "backgroundColor": "#ffffff"
-    },
-    "updates": {
-      "fallbackToCacheTimeout": 0
-    },
-    "assetBundlePatterns": [
-      "**/*"
-    ],
-    "ios": {
-      "supportsTablet": true
-    }
+    "sdkVersion": "27.0.0"
   }
 }
