+++ MotorVehicleManagerSystem/tslint.json
@@ -1,0 +1,124 @@
+{
+
+    "rules": {
+      "member-access": false,
+      "member-ordering": [
+          true,
+          "public-before-private",
+          "static-before-instance",
+          "variables-before-functions"
+      ],
+      "no-any": false,
+      "no-inferrable-types": [false],
+      "no-internal-module": true,
+      "no-var-requires": true,
+      "typedef": [false],
+      "typedef-whitespace": [
+        true, {
+          "call-signature": "nospace",
+          "index-signature": "nospace",
+          "parameter": "nospace",
+          "property-declaration": "nospace",
+          "variable-declaration": "nospace"
+        }, {
+          "call-signature": "space",
+          "index-signature": "space",
+          "parameter": "space",
+          "property-declaration": "space",
+          "variable-declaration": "space"
+        }
+      ],
+      "ban": false,
+      "curly": false,
+      "forin": true,
+      "label-position": true,
+      "no-arg": true,
+      "no-bitwise": true,
+      "no-conditional-assignment": true,
+      "no-console": [
+        true,
+        "debug",
+        "info",
+        "time",
+        "timeEnd",
+        "trace"
+      ],
+      "no-construct": true,
+      "no-debugger": true,
+      "no-duplicate-variable": true,
+      "no-empty": true,
+      "no-eval": true,
+      "no-null-keyword": true,
+      "no-shadowed-variable": true,
+      "no-string-literal": true,
+      "no-switch-case-fall-through": true,
+      "no-unused-expression": true,
+      "no-use-before-declare": true,
+      "no-var-keyword": true,
+      "radix": true,
+      "switch-default": true,
+      "triple-equals": [
+        true,
+        "allow-undefined-check"
+      ],
+      "eofline": false,
+      "indent": [
+        true,
+        "spaces"
+      ],
+      "max-line-length": [
+        true,
+        150
+      ],
+      "no-require-imports": false,
+      "no-trailing-whitespace": true,
+      "object-literal-sort-keys": false,
+      "trailing-comma": [
+        true, {
+          "multiline":  "never",
+          "singleline": "never"
+        }
+      ],
+      "align": [true],
+      "class-name": true,
+      "comment-format": [
+        true,
+        "check-space"
+      ],
+      "interface-name": [false],
+      "jsdoc-format": true,
+      "no-consecutive-blank-lines": [true],
+      "no-parameter-properties": false,
+      "one-line": [
+        true,
+        "check-open-brace",
+        "check-catch",
+        "check-else",
+        "check-finally",
+        "check-whitespace"
+      ],
+      "quotemark": [
+        true,
+        "single",
+        "avoid-escape"
+      ],
+      "semicolon": [
+        true,
+        "always",
+        "ignore-interfaces"
+      ],
+      "variable-name": [
+        true,
+        "allow-leading-underscore",
+        "ban-keywords"
+      ],
+      "whitespace": [
+        true,
+        "check-branch",
+        "check-decl",
+        "check-operator",
+        "check-separator",
+        "check-type"
+      ]
+    }
+}
